refactor: 드롭다운 컴포넌트 하나로 병합
