+++ components/dropdown/dropdown.tsx
@@ -1,16 +1,5 @@
 import useClickOutside from "@/hooks/use-click-outside";
-import {
-  ReactNode,
-  createContext,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
-
-import Body from "./body";
-import Button from "./button";
-import Item from "./item";
+import { ReactNode, createContext, useContext, useState } from "react";
 
 interface DropdownState {
   isDropdownOpen: boolean;
@@ -73,3 +62,77 @@ export function Dropdown({ children, defaultSelected }: DropdownProps) {
 Dropdown.Button = Button;
 Dropdown.Body = Body;
 Dropdown.Item = Item;
+
+// NOTE - Button
+
+interface DropdownButtonProps {
+  children: ReactNode;
+}
+
+/**
+ * @author : 김서영
+ * 드롭다운에서 Select field 영역입니다.
+ * Select field는 드롭다운 리스트에서 선택한 항목을 필드에 표시하는 영역입니다.
+ * 클릭 시 드롭다운 리스트가 열립니다.
+ * 항목을 선택하는 경우 선택한 항목이 필드에 표시됩니다.
+ * @param children : 선택 항목과 관련 없는 토글 아이콘 등이 사용됩니다
+ * @returns : 버튼 컴포넌트를 반환합니다.
+ * @example : <Dropdown.Button>▽</Dropdown.Button>
+ **/
+function Button({ children }: DropdownButtonProps) {
+  const { handleDropdown, selected } = useDropdown();
+
+  return (
+    <button onClick={handleDropdown} className="flex w-full justify-between">
+      <div>{selected}</div>
+      {children}
+    </button>
+  );
+}
+
+// NOTE - Body
+interface BodyProps {
+  children: ReactNode;
+  styles: string;
+}
+
+/**
+ * @author 김서영
+ * 클릭 시에 나타나는 드롭다운 리스트를 감싸는 부분입니다
+ * @param children : Item들이 children에 해당됩니다.
+ * @param styles : 너비 및 배경색 등 추가적으로 적용될 스타일을 지정해주는 프롭입니다.
+ * @example  <Dropdown.Body styles="w-36 bg-blue-200">...</Dropdown.Body>
+ **/
+function Body({ children, styles }: BodyProps) {
+  const { isDropdownOpen } = useDropdown();
+
+  return isDropdownOpen ? (
+    <ul className={`${styles} absolute z-50 mt-2`}>{children}</ul>
+  ) : null;
+}
+
+// NOTE - Item
+interface ItemProps {
+  children: ReactNode;
+}
+
+/**
+ * @author 김서영
+ * 드롭다운 선택 항목에 대한 컴포넌트입니다.
+ * @param children : li 안에 포함될 내용을 적습니다
+ * @param styles : 너비 및 배경색 등 추가적으로 적용될 스타일을 지정해주는 프롭입니다.
+ * @example  <Dropdown.Item><div className="flex gap-2"><p>Seo</p><span>Young</span></div></Dropdown.Item>
+ **/
+function Item({ children }: ItemProps) {
+  const { handleSelect, handleDropdown } = useDropdown();
+
+  const onSelect = () => {
+    handleSelect(children);
+    handleDropdown();
+  };
+  return (
+    <li className="cursor-pointer" onClick={onSelect}>
+      {children}
+    </li>
+  );
+}
